+++ packages/engine-cursor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @holdpoint/engine-cursor
 
+## 0.1.0-alpha.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @holdpoint/types@0.1.0-alpha.6
+
 ## 0.1.0-alpha.7
 
 ### Patch Changes